+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Possibility to load backbone weights in LT-DETR.
 - ONNX export for LT-DETR.
-- Pre-trained ViT-T(+) checkpoints.
+- Pretrained ViT-T(+) checkpoints.
 
 ### Changed